+++ roc/compiler/__init__.py
@@ -1,1 +1,0 @@
-# Roc compiler package (front-end and IR)